refactor: fold serve-sim-client UI components into serve-sim package
Move the simulator UI components, avcc-codec, and types that were only
used internally by serve-sim out of the separate serve-sim-client package
and into serve-sim/src/client/. Delete the serve-sim-client package entirely.

- Moved simulator/ directory (11 files) to src/client/simulator/
- Moved avcc-codec.ts to src/client/avcc-codec.ts
- Created src/client/types.ts with StreamConfig, SimulatorOrientation,
  StreamAPI, AdaptiveState, ConnectionQuality
- Updated 7 client files to import from local ./simulator instead of
  serve-sim-client/simulator
- Migrated 6 test files (avcc-codec, digital-crown, scroll-wheel,
  simulator-geometry, SimulatorFrame, SimulatorToolbar) to serve-sim's
  test suite
- Removed serve-sim-client devDependency from serve-sim/package.json
- Removed serve-sim-client entry from knip.json
- Deleted packages/serve-sim-client/
