+++ knip.json
@@ -24,11 +24,6 @@
         "scripts/build-trails.ts"
       ]
     },
-    "packages/serve-sim-client": {
-      "ignoreDependencies": [
-        "react-dom",
-        "@types/react-dom"
-      ]
-    }
+
   }
 }
