+++ packages/serve-sim/package.json
@@ -80,7 +80,6 @@
     "preact": "^10.29.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "serve-sim-client": "workspace:*",
     "tailwindcss": "^4.1.7",
     "typescript": "^5.7.0"
   },
