+++ packages/serve-sim/src/__tests__/SimulatorToolbar.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "bun:test";
 import { renderToStaticMarkup } from "react-dom/server";
-import { SimulatorToolbar, homeButtonCommand } from "../simulator/SimulatorToolbar";
+import { SimulatorToolbar, homeButtonCommand } from "../client/simulator/SimulatorToolbar";
 
 const exec = async () => ({ stdout: "", stderr: "", exitCode: 0 });
