+++ packages/serve-sim/src/client/components/simulator-resize-corner-handle.tsx
@@ -8,7 +8,7 @@ import {
   simulatorResizeCornerArc,
   type DeviceType,
   type StreamConfig,
-} from "serve-sim-client/simulator";
+} from "../simulator";
 import { useSimulatorResize } from "../hooks/use-simulator-resize";
 import { usePrefersMoreContrast } from "../hooks/use-prefers-more-contrast";
 import { usePrefersReducedMotion } from "../hooks/use-prefers-reduced-motion";
